pyproject.toml
[tool.poetry]
name = "tools-for-agents"
version = "0.1.0"
description = "AI-agent-powered backend for managing appointments in service businesses"
authors = ["tools-for-agents"]
packages = [{include = "domain", from = "src"}, {include = "application", from = "src"}, {include = "infrastructure", from = "src"}, {include = "interfaces", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.29"
alembic = "^1.13"
pydantic-settings = "^2.3"
fastapi = {extras = ["standard"], version = "^0.115"}
openai = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-mock = "^3.14"
pytest-cov = "^5.0"
factory-boy = "^3.3"
pytest-asyncio = "^1.3.0"
httpx = "^0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
